Code re-organization and refactoring
Remove "I" from file names
Split classes into its own files
Create Functions folder to contain all Azure function classes
Remove "Services" folder and move everything to "Core"
