+++ Api/Functions/Encoder.cs
