+++ Api/Repository/Repository.cs
@@ -1,43 +1,8 @@
-using System.Reflection;
 using Microsoft.Azure.Cosmos.Table;
 using System.Threading.Tasks;
-using System.Diagnostics.Contracts;
 
 namespace NosAyudamos
 {
-    interface IRepositoryFactory
-    {
-        IRepository<T> Create<T>() where T : class, ITableEntity;
-    }
-
-    class RepositoryFactory : IRepositoryFactory
-    {
-        readonly IEnvironment enviroment;
-
-        public RepositoryFactory(IEnvironment enviroment) => this.enviroment = enviroment;
-
-        public IRepository<T> Create<T>()
-            where T : class, ITableEntity
-        {
-            var table = typeof(T).GetTypeInfo().GetCustomAttribute<TableAttribute>();
-
-            Contract.Assert(table != null);
-
-            return new Repository<T>(
-                enviroment.GetVariable("StorageConnectionString"), table?.Name!);
-        }
-    }
-
-    interface IRepository<T>
-        where T : class, ITableEntity
-    {
-        Task<T> PutAsync(T entity);
-
-        Task DeleteAsync(T entity);
-
-        Task<T> GetAsync(string partitionKey, string rowKey);
-    }
-
     class Repository<T> : IRepository<T>
         where T : class, ITableEntity
     {
@@ -96,5 +61,15 @@ namespace NosAyudamos
 
             return table;
         }
+    }
+
+    interface IRepository<T>
+        where T : class, ITableEntity
+    {
+        Task<T> PutAsync(T entity);
+
+        Task DeleteAsync(T entity);
+
+        Task<T> GetAsync(string partitionKey, string rowKey);
     }
 }
