+++ Api/Workflow/Workflow.cs
@@ -1,0 +1,15 @@
+using System.Threading.Tasks;
+
+namespace NosAyudamos
+{
+    interface IWorkflow
+    {
+        Task RunAsync(Message message);
+    }
+
+    enum Workflow
+    {
+        Donor,
+        Donee
+    }
+}
